+++ paper/paper.md
@@ -10,7 +10,7 @@ tags:
   
 authors:
   - name: Ole Bialas
-    orcid: 0000-0000-0000-0000
+    orcid: 0000-0003-4472-7626
     equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
     affiliation: 1
     corresponding: true # (This is how to denote the corresponding author)
