+++ paper/paper.bib
@@ -4,6 +4,7 @@
   journal={BioRxiv},
   pages={2021--08},
   year={2021},
+  doi = {10.1101/2021.08.01.454687},
   publisher={Cold Spring Harbor Laboratory}
 }
 
@@ -11,7 +12,8 @@
   title={naplib-python: Neural Acoustic Data Processing and Analysis Tools in Python},
   author={Mischler, Gavin and Raghavan, Vinay and Keshishian, Menoua and Mesgarani, Nima},
   journal={arXiv preprint arXiv:2304.01799},
-  year={2023}
+  year={2023},
+  doi = {10.48550/arXiv.2304.01799}
 }
 
 @article{hamilton2020,
@@ -22,6 +24,7 @@
   number={5},
   pages={573--582},
   year={2020},
+  doi = {10.1080/23273798.2018.1499946},
   publisher={Taylor \& Francis}
 }
 
@@ -32,6 +35,7 @@
   volume={10},
   pages={604},
   year={2016},
+  doi = {10.3389/fnhum.2016.00604},
   publisher={Frontiers Media SA}
 }
 
@@ -43,6 +47,7 @@
   number={7},
   pages={1697--1706},
   year={2015},
+  doi={10.1093/cercor/bht355},
   publisher={Oxford University Press}
 }
 
@@ -53,6 +58,7 @@
   volume={87},
   pages={96--110},
   year={2014},
+  doi={10.1016/j.neuroimage.2013.10.067}
   publisher={Elsevier}
 }
 
@@ -64,6 +70,7 @@
   number={19},
   pages={2457--2465},
   year={2015},
+  doi={10.1016/j.cub.2015.08.030}
   publisher={Elsevier}
 }
 
@@ -75,6 +82,7 @@
   number={5},
   pages={803--809},
   year={2018},
+  doi={10.1016/j.cub.2018.01.080},
   publisher={Elsevier}
 }
 
@@ -84,5 +92,6 @@
   journal={Frontiers in neuroscience},
   pages={267},
   year={2013},
+  doi={10.3389/fnins.2013.00267},
   publisher={Frontiers}
 }
