+++ paper/paper.bib
@@ -58,7 +58,7 @@
   volume={87},
   pages={96--110},
   year={2014},
-  doi={10.1016/j.neuroimage.2013.10.067}
+  doi={10.1016/j.neuroimage.2013.10.067},
   publisher={Elsevier}
 }
 
@@ -70,7 +70,7 @@
   number={19},
   pages={2457--2465},
   year={2015},
-  doi={10.1016/j.cub.2015.08.030}
+  doi={10.1016/j.cub.2015.08.030},
   publisher={Elsevier}
 }
 
