### 2023-05-22 #5484 @OleBialas
> @editorialbot query scope
>
> @OleBialas – due to the relatively small size of this code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next week. If you want to fix the DOIs (noting that @editorialbot's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
>
> @editorialbot generate pdf @editorialbot check references
>
> Finally, the paper isn't compiling properly as you've entered an erroneous ORCID:
>
> > Problem with ORCID (0000-0000-0000-0000) for Ole Bialas. Invalid ORCID (Theoj::Error)

Thank you @arfon! We are aware of the submission guidelines - we discussed waiting with the submission but decided that the researchers who are already using the package would benefit from a timely review and publication of the software.
