### 2023-05-23 266bb01
Fix for predicting a backward model after training without error, with test code
