### 2023-06-09 33b42bb
fix banded ridge for >2 bands
