### 2023-07-14 #5657 @editorialbot
[REVIEW]: mTRFpy: A Python package for temporal response function analysis
**Submitting author:** @OleBialas (<a href="http://orcid.org/0000-0003-4472-7626">Ole Bialas</a>)
**Repository:** https://github.com/powerfulbean/mTRFpy
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.2
