### 2023-07-14 f35eb1b
fix mne_info
