### 2023-07-28 8b8f2a4
TRF.test with one lambda
