### 2023-07-28 dd6fea2
debug test for TRF.test
