### 2023-08-03 e9ca903
minor typo fixes paper
