+++ setup.py
@@ -4,12 +4,12 @@ import re
 with open("README.md") as f:
     readme = f.read()
 
-    # extract version
-    with open("mtrf/__init__.py") as file:
-        for line in file.readlines():
-            m = re.match("__version__ *= *['\"](.*)['\"]", line)
-            if m:
-                version = m.group(1)
+# extract version
+with open("mtrf/__init__.py") as file:
+    for line in file.readlines():
+        m = re.match("__version__ *= *['\"](.*)['\"]", line)
+        if m:
+            version = m.group(1)
 
 setup(
     name="mtrf",
