### 2023-08-03 powerfulbean/mTRFpy#10 @sappelhoff
[JOSS] consolidate doc requirement files
- xref: https://github.com/openjournals/joss-reviews/issues/5657

IMHO, these 3 files should just be one file, called `docs/requirements.txt`:

- https://github.com/powerfulbean/mTRFpy/blob/0df8aa66c23e480829336889b977ff7bff9e1c54/docs/requirements.in#L1
- https://github.com/powerfulbean/mTRFpy/blob/0df8aa66c23e480829336889b977ff7bff9e1c54/requirements_rtd.txt#L1
- https://github.com/powerfulbean/mTRFpy/blob/0df8aa66c23e480829336889b977ff7bff9e1c54/docs/requirements.txt#L1
