+++ .github/workflows/python-package.yml
@@ -27,7 +27,6 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install requests
         pip install -e .[testing]
     - name: Lint with flake8
       run: |
