+++ setup.py
@@ -27,6 +27,10 @@ setup(
         "testing": [
             "requests", "flake8", "pytest", "tqdm", "matplotlib", "scipy",
             ],
+        "docs": [
+            "sphinx", "sphinx_rtd_theme", "mne", "matplotlib"
+        ],
+        "full": ["mtrf[testing]", "mtrf[docs]"]
         },
     packages=find_packages(),
     zip_safe=False,
