+++ docs/conf.py
@@ -62,11 +62,6 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 #
 html_theme = "sphinx_rtd_theme"
 
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ["_static"]
-
 napoleon_numpy_docstring = True
 napoleon_include_init_with_doc = True
 napoleon_use_ivar = True
