### 2023-08-03 powerfulbean/mTRFpy#12 @sappelhoff
[JOSS] clean up branches
- xref: https://github.com/openjournals/joss-reviews/issues/5657

delete old ones: https://github.com/powerfulbean/mTRFpy/branches and unneeded branches
