### 2023-08-03 powerfulbean/mTRFpy#13 @sappelhoff
[JOSS] additional docs
- xref: https://github.com/openjournals/joss-reviews/issues/5657

- [x] please describe the release process of the software, e.g., in the contributor documentation -- potentially listing maintainers.
- [x] please add how to setup up a developer environment / installation for contributors
- [x] please add how to run the test suite, how to format the code (black), and how to build the docs locally (see also some changes in #9)
