+++ paper/paper.md
@@ -18,6 +18,7 @@ authors:
     equal-contrib: true
     affiliation: 1 # (Multiple affiliations must be quoted)
   - name: Edmund C. Lalor
+    orcid: 0000-0002-2498-6631
     affiliation: "1, 2"
 affiliations:
  - name: Department of Biomedical Engineering, University of Rochester, USA
@@ -29,26 +30,47 @@ bibliography: paper.bib
 ---
 
 # Summary
-Traditionally, studies on the neural processing of speech involved the repetitive display of isolated tokens (e.g. phonemes, words, sentences) where the properties of interest were carefully controlled. Recently, researchers have increasingly focused on  investigating brain responses to more naturalistic speech like audiobooks [@hamilton2020]. However, this approach demands statistical tools to account for the different sources of variance that naturally occur in speech. Among the most popular tools to model neural responses to naturalistic speech are multivariate temporal response functions (mTRFs). 
+Traditionally, studies on the neural processing of speech involved the repetitive display of isolated tokens (e.g. phonemes, words, sentences) where the properties of interest were carefully controlled. 
+Recently, researchers have increasingly focused on  investigating brain responses to more naturalistic speech like audiobooks [@hamilton2020]. 
+However, this approach demands statistical tools to account for the different sources of variance that naturally occur in speech. 
+Among the most popular tools to model neural responses to naturalistic speech are multivariate temporal response functions (mTRFs). 
 
-One of the most commonly used packages for computing mTRFs with regularized regression is the mTRF-toolbox [@crosse2016]. However, this toolbox is implemented in the proprietary MATLAB language, restricting accessibility for parts of the scientific community. To overcome this constraint, we present mTRFpy, a Python package which replicates and advances the functionality of the original mTRF-toolbox.
+One of the most commonly used packages for computing mTRFs with regularized regression is the mTRF-toolbox [@crosse2016]. 
+However, this toolbox is implemented in the proprietary MATLAB language, restricting accessibility for parts of the scientific community. 
+To overcome this constraint, we present mTRFpy, a Python package which replicates and advances the functionality of the original mTRF-toolbox.
 
 # Background
-In a nutshell, the mTRF is a regularized linear regression between two continuous signals, computed across multiple time-delays or lags. This accounts for the fact that the relationship between stimulus and neural response is not instantaneous and that the signals are auto-correlated. 
+In a nutshell, the mTRF is a regularized linear regression between two continuous signals, computed across multiple time-delays or lags. 
+This accounts for the fact that the relationship between stimulus and neural response is not instantaneous and that the signals are auto-correlated. 
 
-mTRFs can be used as forward or encoding models to predict (multiple) univariate brain responses as the weighted sum of various acoustic and linguistic speech features while identifying their relative contributions [@diliberto2015, broderick2018]. In this case the model's weights have a clear physiological interpretation because they denote the expected change in neural response following a unit change in a given predictor [@haufe2014]. Thus, they can be understood as a generalization of the event potential, obtained from averaging responses to repetitions of prototypical stimuli for continuous data.
+mTRFs can be used as forward or encoding models to predict (multiple) univariate brain responses as the weighted sum of various acoustic and linguistic speech features while identifying their relative contributions [@diliberto2015, broderick2018]. 
+In this case the model's weights have a clear physiological interpretation because they denote the expected change in neural response following a unit change in a given predictor [@haufe2014]. 
+Thus, they can be understood as a generalization of the event potential, obtained from averaging responses to repetitions of prototypical stimuli for continuous data.
 
[…]
