+++ mtrf/model.py
@@ -335,7 +335,7 @@ class TRF:
         best_regularization: numpy.ndarray
             Optimal regularization values for k training sets.
         """
-        if average is False:
+        if average is False and not np.isscalar(regularization):
             raise ValueError("Average must be True or a list of indices!")
         stimulus, response, n_trials = _check_data(stimulus, response, min_len=3)
         k = _check_k(k, n_trials)
@@ -353,7 +353,7 @@ class TRF:
         r_test, mse_test = np.zeros(n_splits), np.zeros(n_splits)
         for isplit in range(n_splits):
             idx_test = splits[isplit]
-            idx_train_val = np.concatenate(splits[:isplit] + splits[isplit + 1 :])
+    j       idx_train_val = np.concatenate(splits[:isplit] + splits[isplit + 1 :])
             if not np.isscalar(regularization):
                 mse = np.zeros(len(regularization))
                 for ir in _progressbar(
