+++ mtrf/__init__.py
@@ -1,5 +1,1 @@
-import sys
-
 __version__ = "1.2.1"
-
-sys.path.append("..\\")
