### 2023-08-04 906f3b4
Update python-package.yml
add style check
