+++ .github/workflows/python-package.yml
@@ -11,34 +11,6 @@ on:
 
 jobs:
 
-  style:
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.11"]
-    steps:
-    - uses: actions/checkout@v3
-    - uses: psf/black@stable
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-    - uses: actions/cache@v3
-      with:
-        path: ${{ env.pythonLocation }}
-        key: style-0-${{ env.pythonLocation }}-${{ hashFiles('setup.cfg') }}-${{ hashFiles('test_requirements.txt') }}
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install pydocstyle flake8 check-manifest black
-    - name: Display versions and environment information
-      run: |
-        python --version
-        which python
-    - name: Run style & documentation tests
-      run: make pep
-
   build:
     runs-on: ubuntu-latest
     strategy:
@@ -54,8 +26,15 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install flake8 pytest requests
+        python -m pip install flake8 pytest requests black
         pip install -e .[testing]
+    - name: Style checking
+      run: |
+        # stop the build if there are Python syntax errors or undefined names
+        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        black --check .
     - name: Test with pytest
       run: |
         pytest
