### 2023-08-04 powerfulbean/mTRFpy#11 @OleBialas
I can reproduce this with the newest Python and Numpy versions. It seems the error is just numerical inaccuracy. After restricting the equality test to the first 10 decimals, the error disappeared.

By the way, the matrix multiplication warning seems to be due to some bug in Numpy as well:
https://github.com/numpy/numpy/issues/24067
