### 2023-08-04 powerfulbean/mTRFpy#10 @OleBialas
removed requirements_rdt.txt and requirements.in
