### 2023-08-07 powerfulbean/mTRFpy#17 @sappelhoff
[JOSS] link repository to Zenodo and add CITATION.cff
- xref: https://github.com/openjournals/joss-reviews/issues/5657

You can integrate zenodo and Github: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content

later in the JOSS submission process you will have to make an archive of your software anyhow, see:

https://joss.readthedocs.io/en/latest/submitting.html#the-review-process

> Upon successful completion of the review, authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as [Zenodo](https://zenodo.org/) or [figshare](https://figshare.com/), get a DOI for the archive, and update the review issue thread with the version number and DOI.

I strongly recommend Zenodo for its excellent integration with GitHub.

A CITATION.cff file can be used to control the metadata on zenodo: https://help.zenodo.org/faq/#github

A CITATION.cff file can also be a nice opportunity to give contributors credit.

See also my comments here: https://github.com/powerfulbean/mTRFpy/pull/8#discussion_r1285101002
