### 2023-08-07 #5657 @sappelhoff
I have completed my review of the package. I find it in good order and a helpful addition to the scientific Python ecosystem. I recommend this software and the associated paper for publication in JOSS. I have several minor issues and PRs open for this software that I will continue to work on with the authors, however they may be considered relatively minor and their outcome will not affect my overall judgment of this project.
