### 2023-08-15 eafb3a5
add info for contributors
