### 2023-08-15 1d6dbe6
remove unused imports/variables
