+++ .github/workflows/publish-to-pypi.yml
@@ -4,6 +4,10 @@ on:
   release:
     types: [published]
 
+jobs:
+  build-n-publish:
+    name: Build and publish Python 🐍 distributions 📦 to PyPI and TestPyPI
+    runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
     - name: Set up Python
