### 2023-08-15 6ad0b5d
fix the bug of remove the feature dimension in cal cov_mat
