### 2023-08-17 47b90bc
add back return best_regularization in TRF.test
