### 2023-08-17 f7e08eb
accommodate the change on returning best_regularization in trf.test
