+++ mtrf/stats.py
@@ -135,7 +135,7 @@ def _cross_validate(
             cov_xx_hat, cov_xy_hat = _reduced_covariance_matrices(
                 x_train, y_train, lags, model.zeropad, bias
             )
-        else:   
+        else:
             cov_xx_hat = cov_xx[idx_train].mean(axis=0)
             cov_xy_hat = cov_xy[idx_train].mean(axis=0)
         w = np.matmul(np.linalg.inv(cov_xx_hat + regmat), cov_xy_hat) / (1 / fs)
