### 2023-08-18 9e5a86e
_reduced_covariance_matrices was accidentally removed during solving style issue
