+++ .github/workflows/python-package.yml
@@ -33,7 +33,8 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-        black --check .
+        # stop the build if black finds some code need to be reformatted
+        black --check . --diff
     - name: Test with pytest
       run: |
         pytest --verbose
