### 2023-08-28 b830919
allow tuple input
