+++ mtrf/model.py
@@ -2,9 +2,7 @@ from pathlib import Path
 from itertools import product
 import pickle
 from collections.abc import Iterable
-
 import numpy as np
-
 from mtrf.stats import _cross_validate, _progressbar, _check_k
 from mtrf.matrices import (
     covariance_matrices,
@@ -49,10 +47,11 @@ class TRF:
     method: str
         Regularization method. Can be 'ridge' (default), 'tikhonov' or 'banded'.
         See documentation for a detailed explanation.
-    pre_cal_cov: bool
-        If True (default), pre-compute and store covariance matrices for each trial
-        during cross-validation to save time and consume more memory. If False,
-        re-compute covariance matrices for each trial during cross-validation.
+    preload: bool
+        If True (default), covariance matrices for all trials will be pre-loaded before
+        cross-validation. This makes optimization faster but consumes more memory. If
+        False, the covariance matrices will be computed on each iteration which is slower
+        but memory efficient.
 
     Attributes
     ----------
@@ -65,23 +64,17 @@ class TRF:
     """
 
     def __init__(
-        self,
-        direction=1,
-        kind="multi",
-        zeropad=True,
-        bias=True,
-        method="ridge",
-        pre_cal_cov=True,
+        self, direction=1, kind="multi", zeropad=True, method="ridge", preload=True
     ):
         self.weights = None
-        self.bias = bias
+        self.bias = None
         self.times = None
         self.fs = None
         self.regularization = None
-        if isinstance(pre_cal_cov, bool):
-            self.pre_cal_cov = pre_cal_cov
+        if isinstance(preload, bool):
+            self.preload = preload
         else:
-            raise ValueError("Parameter pre_cal_cov must be either True or False!")
+            raise ValueError("Parameter preload must be either True or False!")
         if direction in [1, -1]:
             self.direction = direction
         else:
@@ -201,23 +194,22 @@ class TRF:
         stimulus, response, n_trials = _check_data(stimulus, response, min_len=min_len)
         if not np.isscalar(regularization):
             k = _check_k(k, n_trials)
-        xs, ys, tmin, tmax = _get_xy(stimulus, response, tmin, tmax, self.direction)
+        x, y, tmin, tmax = _get_xy(stimulus, response, tmin, tmax, self.direction)
         lags = list(range(int(np.floor(tmin * fs)), int(np.ceil(tmax * fs)) + 1))
         if self.method == "banded":
             coefficients = list(product(regularization, repeat=len(bands)))
             regularization = [
-                banded_regularization(len(lags), c, bands, self.bias)
-                for c in coefficients
+                banded_regularization(len(lags), c, bands) for c in coefficients
             ]
         if np.isscalar(regularization):
-            self._train(xs, ys, fs, tmin, tmax, regularization)
+            self._train(x, y, fs, tmin, tmax, regularization)
             return
         else:  # run cross-validation once per regularization parameter
             # pre-compute covariance matrices
             cov_xx, cov_xy = None, None
-            if self.pre_cal_cov:
+            if self.preload:
                 cov_xx, cov_xy = covariance_matrices(
-                    xs, ys, lags, self.zeropad, self.bias
+                    x, y, lags, self.zeropad, self.preload
                 )
             else:
                 cov_xx, cov_xy = None, None
@@ -230,8 +222,8 @@ class TRF:
             ):
                 r[ir], mse[ir] = _cross_validate(
                     self.copy(),
-                    xs,
-                    ys,
+                    x,
[…]
