+++ tests/test_model.py
@@ -8,22 +8,47 @@ n = np.random.randint(3, 10)
 stimulus, response, fs = load_sample_data(n_segments=n)
 
 
+def test_loss_function():
+    def mean_absolute_deviation(y, y_pred):
+        return np.abs(y - y_pred).mean()
+
+    def mean_difference(y, y_pred):
+        return np.mean(y - y_pred)
+
+    for loss_function in [mean_absolute_deviation, mean_difference]:
+        trf = TRF(loss_function=mad)
+        regularization = [
+            np.random.uniform(0, 1000) for _ in range(np.random.randint(2, 5))
+        ]
+        tmin = np.random.uniform(-0.1, 0.05)
+        tmax = np.random.uniform(0.1, 0.4)
+        loss = trf.train(stimulus, response, fs, tmin, tmax, regularization)
+        assert trf.regularization == regularization[np.argmin(loss)]
+
+
 def test_train():
     tmin = np.random.uniform(-0.1, 0.05)
     tmax = np.random.uniform(0.1, 0.4)
-    direction = np.random.choice([1, -1])
     regularization = np.random.uniform(0, 10)
-    trf = TRF(direction=direction)
-    trf.train(stimulus, response, fs, tmin, tmax, regularization)
-    if direction == 1:
-        assert trf.weights.shape[0] == stimulus[0].shape[-1]
-        assert trf.weights.shape[-1] == response[0].shape[-1]
-    if direction == -1:
-        assert trf.weights.shape[-1] == stimulus[0].shape[-1]
-        assert trf.weights.shape[0] == response[0].shape[-1]
-    regularization = [np.random.uniform(0, 10) for _ in range(randint(2, 5))]
-    r, mse = trf.train(stimulus, response, fs, tmin, tmax, regularization)
-    assert len(r) == len(mse) == len(regularization)
+    for direction in [1, -1]:
+        trf = TRF(direction=direction)
+        trf.train(stimulus, response, fs, tmin, tmax, regularization)
+        if direction == 1:
+            assert trf.weights.shape[0] == stimulus[0].shape[-1]
+            assert trf.weights.shape[-1] == response[0].shape[-1]
+        if direction == -1:
+            assert trf.weights.shape[-1] == stimulus[0].shape[-1]
+            assert trf.weights.shape[0] == response[0].shape[-1]
+
+
+def test_optimize():
+    for direction in [1, -1]:
+        tmin = np.random.uniform(-0.1, 0.05)
+        tmax = np.random.uniform(0.1, 0.4)
+        trf = TRF(direction=direction)
+        regularization = [np.random.uniform(0, 10) for _ in range(randint(2, 5))]
+        loss = trf.train(stimulus, response, fs, tmin, tmax, regularization)
+    assert len(loss) == len(regularization)
 
 
 def test_predict():
@@ -33,12 +58,12 @@ def test_predict():
     trf = TRF()
     trf.train(stimulus, response, fs, tmin, tmax, regularization)
     for average in [True, list(range(randint(response[0].shape[-1])))]:
-        prediction, r, mse = trf.predict(stimulus, response, average=average)
+        prediction, loss = trf.predict(stimulus, response, average=average)
         assert len(prediction) == len(response)
         assert all([p[0].shape == r[0].shape for p, r in zip(prediction, response)])
-        assert np.isscalar(r) and np.isscalar(mse)
-    prediction, r, mse = trf.predict(stimulus, response, average=False)
-    assert r.shape[-1] == mse.shape[-1] == trf.weights.shape[-1]
+        assert np.isscalar(loss)
[…]
