+++ mtrf/model.py
@@ -3,7 +3,13 @@ from itertools import product
 import pickle
 from collections.abc import Iterable
 import numpy as np
-from mtrf.stats import _cross_validate, _progressbar, _check_k
+from mtrf.stats import (
+    _cross_validate,
+    _progressbar,
+    _check_k,
+    mean_squared_error,
+    one_minus_r,
+)
 from mtrf.matrices import (
     covariance_matrices,
     banded_regularization,
@@ -64,13 +70,23 @@ class TRF:
     """
 
     def __init__(
-        self, direction=1, kind="multi", zeropad=True, method="ridge", preload=True
+        self,
+        direction=1,
+        kind="multi",
+        zeropad=True,
+        method="ridge",
+        preload=True,
+        loss_function=one_minus_r,
     ):
         self.weights = None
         self.bias = None
         self.times = None
         self.fs = None
         self.regularization = None
+        if not callable(loss_function):
+            raise ValueError("Loss function must be callable")
+        else:
+            self.loss_function = loss_function
         if isinstance(preload, bool):
             self.preload = preload
         else:
@@ -140,7 +156,7 @@ class TRF:
         Compute a linear mapping between stimulus and response, or vice versa, using
         regularized regression. This method can compare multiple values for the
         regularization parameter and select the best one (i.e. the one that yields
-        the highest prediction accuracy).
+        the lowest prediction loss).
 
         Parameters
         ----------
@@ -170,20 +186,20 @@ class TRF:
             Number of data splits for cross validation, defaults to 5.
             If -1, do leave-one-out cross-validation.
         average: bool or list or numpy.ndarray
-            If True (default), average correlation and mean squared error across all
-            predictions (e.g. channels in the case of forward modelling). If `average`
-            is an array of integers only average the predicted features at those indices.
+            If True (default), average loss cross all predictions (e.g. channels in the
+            case of forward modelling). If `average` is an array of indices only average
+            those features.
         seed: int
             Seed for the random number generator.
         verbose: bool
-            If True (default), show a progress bar during fitting
+            If True (default), show a progress bar during fitting.
 
         Returns
         -------
-        r : list
-            Correlation of prediction and actual output.
-        mse : list
-            Mean squared error between prediction and actual output.
+        loss : list
+            When providing multiple `regularization` values this returns the loss as
+            computed by the loss function defined in the `TRF.loss_function` attribute
+            for every regularization value.
         """
         if average is False:
             raise ValueError("Average must be True or a list of indices!")
@@ -213,14 +229,13 @@ class TRF:
                 )
             else:
                 cov_xx, cov_xy = None, None
-            r = np.zeros(len(regularization))
-            mse = np.zeros(len(regularization))
+            loss = np.zeros(len(regularization))
             for ir in _progressbar(
                 range(len(regularization)),
                 "Hyperparameter optimization",
                 verbose=verbose,
             ):
-                r[ir], mse[ir] = _cross_validate(
+                loss[ir] = _cross_validate(
                     self.copy(),
                     x,
                     y,
@@ -234,9 +249,9 @@ class TRF:
                     average=average,
                     verbose=verbose,
                 )
-            best_regularization = list(regularization)[np.argmin(mse)]
+            best_regularization = list(regularization)[np.argmin(loss)]
             self._train(x, y, fs, tmin, tmax, best_regularization)
-            return r, mse
+            return loss
 
     def _train(self, x, y, fs, tmin, tmax, regularization):
         if isinstance(regularization, np.ndarray):  # check if matrix is diagonal
@@ -325,16 +340,15 @@ class TRF:
         seed: int
             Seed for the random number generator.
         verbose: bool
[…]

[…]
