+++ tests/test_stats.py
@@ -7,20 +7,20 @@ stimulus, response, fs = load_sample_data(n_segments=n)
 
 
 def test_crossval():
-    tmin, tmax = np.random.uniform(-0.1, 0.05), np.random.uniform(0.1, 0.4)
-    direction = np.random.choice([1, -1])
-    reg = np.random.uniform(0, 10)
-    trf = TRF(direction=direction)
-    splits = np.random.randint(2, 5)
-    r, mse = cross_validate(trf, stimulus, response, fs, tmin, tmax, reg, splits)
-    assert np.isscalar(r) and np.isscalar(mse)
-    r, mse = cross_validate(
-        trf, stimulus, response, fs, tmin, tmax, reg, splits, average=False
-    )
-    if direction == 1:
-        assert len(r) == len(mse) == response[0].shape[-1]
-    else:
-        assert len(r) == len(mse) == stimulus[0].shape[-1]
+    for direction in [1, -1]:
+        tmin, tmax = np.random.uniform(-0.1, 0.05), np.random.uniform(0.1, 0.4)
+        reg = np.random.uniform(0, 10)
+        trf = TRF(direction=direction)
+        splits = np.random.randint(2, 5)
+        loss = cross_validate(trf, stimulus, response, fs, tmin, tmax, reg, splits)
+        assert np.isscalar(loss)
+        loss = cross_validate(
+            trf, stimulus, response, fs, tmin, tmax, reg, splits, average=False
+        )
+        if direction == 1:
+            assert len(loss) == response[0].shape[-1]
+        else:
+            assert len(loss) == stimulus[0].shape[-1]
 
 
 def test_permutation():
@@ -28,7 +28,7 @@ def test_permutation():
     n_permute = np.random.randint(5, 100)
     reg = np.random.uniform(0, 10)
     trf = TRF()
-    r, mse = permutation_distribution(
+    loss = permutation_distribution(
         trf,
         stimulus,
         response,
@@ -40,4 +40,4 @@ def test_permutation():
         k=-1,
         average=[1, 2, 3],
     )
-    assert len(r) == len(mse) == n_permute
+    assert len(loss) == n_permute
