### 2023-08-31 59d998a
rename loss with metric, remove test argument in train, and move if else outside loop
