+++ mtrf/stats.py
@@ -9,14 +9,9 @@ from mtrf.matrices import (
 )
 
 
-def mean_squared_error(y, y_pred):
-    mse = np.mean((y - y_pred) ** 2, axis=0)
-    return mse
-
-
-def one_minus_r(y, y_pred):
+def neg_mean_squared_error(y, y_pred):
     """
-    Compute one minus the Pearson's correlation coefficient between predicted
+    Compute negative mean suqare error (mse) between predicted
     and observed data
 
     Parameters
@@ -28,13 +23,34 @@ def one_minus_r(y, y_pred):
 
     Returns
     -------
-    inv_r: np.ndarray
-        Inverse Pearsons r (1-r) for each feature in y.
+    neg_mse: np.ndarray
+        Negative mse (-mse) for each feature in y.
+    """
+    mse = np.mean((y - y_pred) ** 2, axis=0)
+    return -mse
+
+
+def pearsonr(y, y_pred):
+    """
+    Compute Pearson's correlation coefficient between predicted
+    and observed data
+
+    Parameters
+    ----------
+    y: np.ndarray
+        samples-by-features matrix of observed data.
+    y_pred: np.ndarray
+        samples-by-features matrix of predicted data.
+
+    Returns
+    -------
+    r: np.ndarray
+        Pearsons r for each feature in y.
     """
     r = np.mean((y - y.mean(0)) * (y_pred - y_pred.mean(0)), 0) / (
         y.std(0) * y_pred.std(0)
     )
-    return 1 - r
+    return r
 
 
 def cross_validate(
@@ -51,7 +67,7 @@ def cross_validate(
     verbose=True,
 ):
     """
-    Test model loss using k-fold cross-validation.
+    Test model metric using k-fold cross-validation.
 
     Input data is randomly shuffled and separated into k parts of with approximately
     the same number of trials. The first k-1 parts are used for training and the kth
@@ -84,12 +100,12 @@ def cross_validate(
         If True (default), average correlation and mean squared error across all
         predictions (e.g. channels in the case of forward modelling). If `average`
         is an array of integers only average the predicted features at those indices.
-        If `False`, return each predicted feature's loss.
+        If `False`, return each predicted feature's metric.
 
     Returns
     -------
-    loss: float or numpy.ndarray
-        Loss as computed by the loss function in  the attribute `model.loss_function`.
+    metric: float or numpy.ndarray
+        Metric as computed by the metric function in the attribute `model.metric_function`.
     """
     trf = model.copy()
     trf.bias, trf.weights = None, None
@@ -100,7 +116,7 @@ def cross_validate(
     x, y, tmin, tmax = _get_xy(stimulus, response, tmin, tmax, model.direction)
     lags = list(range(int(np.floor(tmin * fs)), int(np.ceil(tmax * fs)) + 1))
     cov_xx, cov_xy = covariance_matrices(x, y, lags, model.zeropad, trf.preload)
-    loss = _cross_validate(
+    metric = _cross_validate(
         model,
         x,
         y,
@@ -113,7 +129,7 @@ def cross_validate(
         average,
         verbose,
     )
-    return loss
+    return metric
 
 
 def _cross_validate(
@@ -144,9 +160,9 @@ def _cross_validate(
     splits = np.array_split(splits, k)
 
     if average is True:
-        loss = np.zeros(k)
+        metric = np.zeros(k)
     else:
-        loss = np.zeros((k, y[0].shape[-1]))
+        metric = np.zeros((k, y[0].shape[-1]))
 
     for isplit in _progressbar(range(len(splits)), "Cross-validating", verbose=verbose):
         idx_val = splits[isplit]
@@ -171,11 +187,11 @@ def _cross_validate(
[…]
