+++ tests/test_model.py
@@ -8,22 +8,22 @@ n = np.random.randint(3, 10)
 stimulus, response, fs = load_sample_data(n_segments=n)
 
 
-def test_loss_function():
-    def mean_absolute_deviation(y, y_pred):
-        return np.abs(y - y_pred).mean()
+def test_metric_function():
+    def neg_mean_absolute_deviation(y, y_pred):
+        return -np.abs(y - y_pred).mean()
 
-    def mean_difference(y, y_pred):
-        return np.mean(y - y_pred)
+    def neg_mean_difference(y, y_pred):
+        return -np.mean(y - y_pred)
 
-    for loss_function in [mean_absolute_deviation, mean_difference]:
-        trf = TRF(loss_function=loss_function)
+    for metric_function in [neg_mean_absolute_deviation, neg_mean_difference]:
+        trf = TRF(metric_function=metric_function)
         regularization = [
             np.random.uniform(0, 1000) for _ in range(np.random.randint(2, 5))
         ]
         tmin = np.random.uniform(-0.1, 0.05)
         tmax = np.random.uniform(0.1, 0.4)
-        loss = trf.train(stimulus, response, fs, tmin, tmax, regularization)
-        assert trf.regularization == regularization[np.argmin(loss)]
+        metric = trf.train(stimulus, response, fs, tmin, tmax, regularization)
+        assert trf.regularization == regularization[np.argmax(metric)]
 
 
 def test_train():
@@ -47,8 +47,8 @@ def test_optimize():
         tmax = np.random.uniform(0.1, 0.4)
         trf = TRF(direction=direction)
         regularization = [np.random.uniform(0, 10) for _ in range(randint(2, 5))]
-        loss = trf.train(stimulus, response, fs, tmin, tmax, regularization)
-    assert len(loss) == len(regularization)
+        metric = trf.train(stimulus, response, fs, tmin, tmax, regularization)
+    assert len(metric) == len(regularization)
 
 
 def test_predict():
@@ -58,21 +58,21 @@ def test_predict():
     trf = TRF()
     trf.train(stimulus, response, fs, tmin, tmax, regularization)
     for average in [True, list(range(randint(response[0].shape[-1])))]:
-        prediction, loss = trf.predict(stimulus, response, average=average)
+        prediction, metric = trf.predict(stimulus, response, average=average)
         assert len(prediction) == len(response)
         assert all([p[0].shape == r[0].shape for p, r in zip(prediction, response)])
-        assert np.isscalar(loss)
-    prediction, loss = trf.predict(stimulus, response, average=False)
-    assert loss.shape[-1] == trf.weights.shape[-1]
+        assert np.isscalar(metric)
+    prediction, metric = trf.predict(stimulus, response, average=False)
+    assert metric.shape[-1] == trf.weights.shape[-1]
     trf = TRF(direction=-1)
     trf.train(stimulus, response, fs, tmin, tmax, regularization)
     for average in [True, list(range(randint(stimulus[0].shape[-1])))]:
-        prediction, loss = trf.predict(stimulus, response, average=average)
+        prediction, metric = trf.predict(stimulus, response, average=average)
         assert len(prediction) == len(stimulus)
         assert all([p[0].shape == s[0].shape for p, s in zip(prediction, stimulus)])
-        assert np.isscalar(loss)
-    prediction, loss = trf.predict(stimulus, response, average=False)
-    assert loss.shape[-1] == trf.weights.shape[-1]
+        assert np.isscalar(metric)
+    prediction, metric = trf.predict(stimulus, response, average=False)
+    assert metric.shape[-1] == trf.weights.shape[-1]
 
 
 def test_test():
[…]

[…]
