### 2023-09-01 a7e4338
fixes to bib file
