+++ paper/paper.bib
@@ -9,7 +9,7 @@
 }
 
 @article{mischler2023,
-  title={naplib-python: Neural Acoustic Data Processing and Analysis Tools in Python},
+  title={{n}aplib-python: Neural Acoustic Data Processing and Analysis Tools in {P}ython},
   author={Mischler, Gavin and Raghavan, Vinay and Keshishian, Menoua and Mesgarani, Nima},
   journal={arXiv preprint arXiv:2304.01799},
   year={2023},
@@ -41,7 +41,7 @@
 
 @article{osullivan2015,
   title={Attentional selection in a cocktail party environment can be decoded from single-trial EEG},
-  author={O'sullivan, James A and Power, Alan J and Mesgarani, Nima and Rajaram, Siddharth and Foxe, John J and Shinn-Cunningham, Barbara G and Slaney, Malcolm and Shamma, Shihab A and Lalor, Edmund C},
+  author={O'Sullivan, James A and Power, Alan J and Mesgarani, Nima and Rajaram, Siddharth and Foxe, John J and Shinn-Cunningham, Barbara G and Slaney, Malcolm and Shamma, Shihab A and Lalor, Edmund C},
   journal={Cerebral cortex},
   volume={25},
   number={7},
@@ -64,7 +64,7 @@
 
 @article{diliberto2015,
   title={Low-frequency cortical entrainment to speech reflects phoneme-level processing},
-  author={Di Liberto, Giovanni M and O’sullivan, James A and Lalor, Edmund C},
+  author={Di Liberto, Giovanni M and O’Sullivan, James A and Lalor, Edmund C},
   journal={Current Biology},
   volume={25},
   number={19},
@@ -88,8 +88,8 @@
 
 @article{gramfort2013,
   title={MEG and EEG data analysis with MNE-Python},
-  author={Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A and Strohmeier, Daniel and Brodbeck, Christian and Goj, Roman and Jas, Mainak and Brooks, Teon and Parkkonen, Lauri and others},
-  journal={Frontiers in neuroscience},
+  author={Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A and Strohmeier, Daniel and Brodbeck, Christian and Goj, Roman and Jas, Mainak and Brooks, Teon and Parkkonen, Lauri and Hämäläinen, Matti},
+  journal={Frontiers in Neuroscience},
   pages={267},
   year={2013},
   doi={10.3389/fnins.2013.00267},
