+++ paper/paper.bib
@@ -87,7 +87,7 @@
 }
 
 @article{gramfort2013,
-  title={MEG and EEG data analysis with MNE-Python},
+  title={MEG and EEG data analysis with MNE-{P}ython},
   author={Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A and Strohmeier, Daniel and Brodbeck, Christian and Goj, Roman and Jas, Mainak and Brooks, Teon and Parkkonen, Lauri and Hämäläinen, Matti},
   journal={Frontiers in Neuroscience},
   pages={267},
