+++ mtrf/stats.py
@@ -9,7 +9,7 @@ from mtrf.matrices import (
 )
 
 
-def neg_mean_squared_error(y, y_pred):
+def neg_mse(y, y_pred):
     """
     Compute negative mean suqare error (mse) between predicted
     and observed data
@@ -53,14 +53,14 @@ def pearsonr(y, y_pred):
     return r
 
 
-def cross_validate(
+def crossval(
     model,
     stimulus,
     response,
-    fs=None,
-    tmin=None,
-    tmax=None,
-    regularization=None,
+    fs,
+    tmin,
+    tmax,
+    regularization,
     k=-1,
     seed=None,
     average=True,
@@ -105,18 +105,16 @@ def cross_validate(
     Returns
     -------
     metric: float or numpy.ndarray
-        Metric as computed by the metric function in the attribute `model.metric_function`.
+        Metric as computed by the metric function in the attribute `model.metric`.
     """
     trf = model.copy()
-    trf.bias, trf.weights = None, None
-    fs, tmin, tmax, regularization = _check_attr(trf, fs, tmin, tmax, regularization)
     if seed is not None:
         random.seed(seed)
     stimulus, response, _ = _check_data(stimulus, response, min_len=2)
     x, y, tmin, tmax = _get_xy(stimulus, response, tmin, tmax, model.direction)
     lags = list(range(int(np.floor(tmin * fs)), int(np.ceil(tmax * fs)) + 1))
     cov_xx, cov_xy = covariance_matrices(x, y, lags, model.zeropad, trf.preload)
-    metric = _cross_validate(
+    metric = _crossval(
         model,
         x,
         y,
@@ -132,7 +130,146 @@ def cross_validate(
     return metric
 
 
-def _cross_validate(
+def nested_crossval(
+    model,
+    stimulus,
+    response,
+    fs,
+    tmin,
+    tmax,
+    regularization,
+    bands=None,
+    k=-1,
+    average=True,
+    seed=None,
+    verbose=True,
+):
+    """
+    Unbiased estimate of model accuracy when fitting the regularization parameter.
+
+    This fuction divides the data into k parts and runs two nested
+    cross-validation loops: the outer loop selects k-1 parts to optimize the
+    regularization value and the kth part to test the final model's accuracy.
+    The inner loop uses cross-validation to determine the best regularization
+    value as in the `fit` method. The data are rotated so that each of the
+    k segments is used to test the final model's accuracy once. The average
+    correlation and mean squared error across all folds is an unbiased estimate
+    of the model's accuracy because the test data was not part of the optimization
+    process.
+
+    Parameters
+    ----------
+    stimulus: list
+        Each element must contain one trial's stimulus in a two-dimensional
+        samples-by-features array (second dimension can be omitted if there is
+        only a single feature.
+    response: list
+        Each element must contain one trial's response in a two-dimensional
+        samples-by-channels array.
+    fs: int
+        Sample rate of stimulus and response in hertz.
+    tmin: float
+        Minimum time lag in seconds.
+    tmax: float
+        Maximum time lag in seconds.
+    regularization: list or float or int
+        Values for the regularization parameter lambda. The model is fitted
+        separately for each value and the one yielding the highest accuracy
+        is chosen (correlation and mean squared error of each model are returned).
+    bands: list or None
+        Must only be provided when using banded ridge regression. Size of the
+        features for which a regularization parameter is fitted, in the order they
+        appear in the stimulus matrix. For example, when the stimulus consists of
+        an envelope vector and a 16-band spectrogram, bands would be [1, 16].
+    k: int
+        Number of data splits for cross validation, defaults to 5.
+        If -1, do leave-one-out cross-validation.
+    average: bool or list or numpy.ndarray
[…]
