+++ mtrf/model.py
@@ -4,10 +4,10 @@ import pickle
 from collections.abc import Iterable
 import numpy as np
 from mtrf.stats import (
-    _cross_validate,
+    _crossval,
     _progressbar,
     _check_k,
-    neg_mean_squared_error,
+    neg_mse,
     pearsonr,
 )
 from mtrf.matrices import (
@@ -58,7 +58,7 @@ class TRF:
         cross-validation. This makes optimization faster but consumes more memory. If
         False, the covariance matrices will be computed on each iteration which is slower
         but memory efficient.
-    metric_function: callable
+    metric: callable
         A callable which accept two arguments (true y, predicted y), and retrun a single
         value for each feature in y. The default is mtrf.stats.pearsonr.
 
@@ -79,17 +79,17 @@ class TRF:
         zeropad=True,
         method="ridge",
         preload=True,
-        metric_function=pearsonr,
+        metric=pearsonr,
     ):
         self.weights = None
         self.bias = None
         self.times = None
         self.fs = None
         self.regularization = None
-        if not callable(metric_function):
+        if not callable(metric):
             raise ValueError("Metric function must be callable")
         else:
-            self.metric_function = metric_function
+            self.metric = metric
         if isinstance(preload, bool):
             self.preload = preload
         else:
@@ -199,7 +199,7 @@ class TRF:
         -------
         metric : list
             When providing multiple `regularization` values this returns the metric as
-            computed by the metric function defined in the `TRF.metric_function` attribute
+            computed by the metric function defined in the `TRF.metric` attribute
             for every regularization value.
         """
         if average is False:
@@ -236,7 +236,7 @@ class TRF:
                 "Hyperparameter optimization",
                 verbose=verbose,
             ):
-                metric[ir] = _cross_validate(
+                metric[ir] = _crossval(
                     self.copy(),
                     x,
                     y,
@@ -280,144 +280,6 @@ class TRF:
         self.times = np.array(lags) / fs
         self.fs = fs
 
-    def test(
-        self,
-        stimulus,
-        response,
-        fs,
-        tmin,
-        tmax,
-        regularization,
-        bands=None,
-        k=-1,
-        average=True,
-        seed=None,
-        verbose=True,
-    ):
-        """
-        Unbiased estimate of model accuracy when fitting the regularization parameter.
-
-        This fuction divides the data into k parts and runs two nested
-        cross-validation loops: the outer loop selects k-1 parts to optimize the
-        regularization value and the kth part to test the final model's accuracy.
-        The inner loop uses cross-validation to determine the best regularization
-        value as in the `fit` method. The data are rotated so that each of the
-        k segments is used to test the final model's accuracy once. The average
-        correlation and mean squared error across all folds is an unbiased estimate
-        of the model's accuracy because the test data was not part of the optimization
-        process.
-
-        Parameters
-        ----------
-        stimulus: list
-            Each element must contain one trial's stimulus in a two-dimensional
-            samples-by-features array (second dimension can be omitted if there is
-            only a single feature.
-        response: list
-            Each element must contain one trial's response in a two-dimensional
-            samples-by-channels array.
-        fs: int
-            Sample rate of stimulus and response in hertz.
-        tmin: float
-            Minimum time lag in seconds.
-        tmax: float
-            Maximum time lag in seconds.
-        regularization: list or float or int
-            Values for the regularization parameter lambda. The model is fitted
-            separately for each value and the one yielding the highest accuracy
-            is chosen (correlation and mean squared error of each model are returned).
-        bands: list or None
-            Must only be provided when using banded ridge regression. Size of the
-            features for which a regularization parameter is fitted, in the order they
[…]

[…]
