+++ mtrf/__init__.py
@@ -1,1 +1,3 @@
 __version__ = "1.2.3"
+
+from mtrf.model import TRF, load_sample_data
