+++ docs/optimization.rst
@@ -1,3 +1,5 @@
+.. currentmodule:: mtrf
+
 Regularization
 ==============
 
@@ -23,10 +25,10 @@ Optimization
 
 We compute multiple models with different values for :math:`\lambda` and select the model that makes the best prediction. 
 This requires a metric for evaluating the model's prediction. 
-We estimate the models accuracy for each candidate value of :math:`lambda` using cross-validation, pick the one that maximizes our metric, and use it to fit a model on the whole data. 
-This whole procedure can be done by using the `TRF.train` function and passing a list instead of a single value for the :py:const"`regularization` parameter. 
-:py:meth:`TF.train` will return the evaluation metric for each candidate value of :math:`\lambda`. 
-In the below example we are optimizing lambda using Pearson's correlation as well as the (negative) mean squared error as metrics and visualize how they change as a function of :math:`lambda`.::
+We estimate the models accuracy for each candidate value of :math:`\lambda` using cross-validation, pick the one that maximizes our metric, and use it to fit a model on the whole data. 
+This whole procedure can be done by using the :py:meth:`TRF.train` function and passing a list instead of a single value for the ``regularization`` parameter. 
+:py:meth:`TRF.train` will return the evaluation metric for each candidate value of :math:`\lambda`. 
+In the below example we are optimizing lambda using Pearson's correlation as well as the (negative) mean squared error as metrics and visualize how they change as a function of :math:`\lambda`.::
 
     import numpy as np
     from matplotlib import pyplot as plt
@@ -73,14 +75,14 @@ On the other hand, picking a :math:`\lambda` value that is too low drastically r
 More overfitting
 ----------------
 In the previous section, we explained how cross-validation can be used to avoid overfitting. However, optimizing the value of :math:`\lambda` creates another source of overfitting. 
-This is because, while the weights are only optimized for the training split, selecting the best :math:`lambda` involves the whole data set. 
-Thus, optimizing :math:`lambda` on the same data used to validate the model inflates its accuracy estimate. To avoid this, we must use two nested cross-validation loops. 
+This is because, while the weights are only optimized for the training split, selecting the best :math:`\lambda` involves the whole data set. 
+Thus, optimizing :math:`\lambda` on the same data used to validate the model inflates its accuracy estimate. To avoid this, we must use two nested cross-validation loops. 
 The outer loop splits the data into a test and a train/validate set. 
-The latter is used to find the best value for :math"`\lambda` and the test set is used to evaluate the final model. This way, the data used to determine the model's accuracy is never part of the optimization process.
-The data are rotated in the outer loop, so that each segment is used for testing once and for each iteration of the outer loop, a full cross-validation loop is performed using the remaing :py:const:`k`-1 segments to find the best value for :math:`\lambda`.
+The latter is used to find the best value for :math:`\lambda` and the test set is used to evaluate the final model. This way, the data used to determine the model's accuracy is never part of the optimization process.
+The data are rotated in the outer loop, so that each segment is used for testing once and for each iteration of the outer loop, a full cross-validation loop is performed using the remaing ``k``-1 segments to find the best value for :math:`\lambda`.
 This makes the nested cross-validation time consuming but it will produce an unbiased estimate of the model's accuracy while optimizing :math:`\lambda`.
 
-This procedure is implemented in the :py:func:`nested_crossval` function which is part of the :py:module:`stats` module and returns the accuracy metric for each test set as well as the value of :math:`\lambda` selected from each train/validate set. In the examle below, we'll use it to compute and unbiased estimate of the models accuracy and compare it to the highest accuracy obtained during optimization in the previous section::
[…]
