+++ mtrf/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "2.0.0"
+__version__ = "2.0.1"
 
 from mtrf.model import TRF, load_sample_data
