+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{brodbeck2021,
-  title={Eelbrain: A Python toolkit for time-continuous analysis with temporal response functions},
+  title={Eelbrain: A {P}ython toolkit for time-continuous analysis with temporal response functions},
   author={Brodbeck, Christian and Das, Proloy and Gillis, Marlies and Kulasingham, Joshua P and Bhattasali, Shohini and Gaston, Phoebe and Resnik, Philip and Simon, Jonathan Z},
   journal={BioRxiv},
   pages={2021--08},
