### 2023-09-06 #5657 @britta-wstnr
Hi @OleBialas,

thanks for starting on the check list.
Could you please:
1. Change the title of the archive to match the title of the paper?
2. Create a release of the software and post the version number?

Here is the link to the full list: https://github.com/openjournals/joss-reviews/issues/5657#issuecomment-1701244325 (Additional Author Tasks)
(You won't be able to tick things off, I will have to do that for you).

Thanks! 🙏
