### 2023-09-06 46eef50
Update CITATION.cff
