+++ CITATION.cff
@@ -2,7 +2,7 @@
 # Visit https://bit.ly/cffinit to generate yours today!
 
 cff-version: 1.2.0
-title: mTRFpy
+title: mTRFpy: A Python package for temporal response function analysis
 message: >-
   If you use this software, please cite it using the
   metadata from this file.
