### 2023-09-06 c6d3978
Update CITATION.cff
