### 2023-09-06 #5657 @powerfulbean
Hi @britta-wstnr,

We just created a latest release and changed the title. Now the latest DOI is: 10.5281/zenodo.8321912.

Thank you!
