### 2023-09-08 #5657 @powerfulbean
> @OleBialas the version number is v2.0.2, is that correct?

Hi @britta-wstnr , yes, the version number is v2.0.2. Really sorry that I forgot to post the version number!
