### 2023-09-12 c13d325
Add citation and JOSS badge
