### 2023-09-12 #5657 @OleBialas
Great, thank you @britta-wstnr! While the DOI resolves, the JOSS website does not show the paper and clicking the download button throws a 404 - did something go wrong with the upload?
