### 2023-09-12 4c16325
Update CITATION.cff joss version
