+++ CITATION.cff
@@ -1,38 +1,44 @@
-# This CITATION.cff file was generated with cffinit.
-# Visit https://bit.ly/cffinit to generate yours today!
-
-cff-version: 1.2.0
-title: 'mTRFpy: A Python package for temporal response function analysis'
-message: >-
-  If you use this software, please cite it using the
-  metadata from this file.
-type: software
+cff-version: "1.2.0"
 authors:
-  - given-names: Ole
-    family-names: Bialas
-    email: ole.bialas@posteo.de
-    affiliation: University of Rochester
-    orcid: 'https://orcid.org/0000-0003-4472-7626'
-  - given-names: Jin
-    family-names: Dou
-    email: jdou3@ur.rochester.edu
-    affiliation: University of Rochester
-    orcid: 'https://orcid.org/0009-0000-0539-5951'
-  - given-names: Edmund
-    family-names: Lalor
-    email: elalor@ur.rochester.edu
-    affiliation: University of Rochester
-    orcid: https://orcid.org/0000-0002-2498-6631
-repository-code: 'https://github.com/powerfulbean/mTRFpy'
-url: 'https://mtrfpy.readthedocs.io/en/latest/'
-abstract: >-
-  mTRFpy is a Python package for computing linear
-  encoding/decoding models, also known multivariate temporal
-  response functions (mTRFs) that are commonly used in
-  neuroscience
-keywords:
-  - Neuroscience
-  - Encoding
-  - Decoding
-  - Modeling
-license: BSD-3-Clause
+- family-names: Bialas
+  given-names: Ole
+  orcid: "https://orcid.org/0000-0003-4472-7626"
+- family-names: Dou
+  given-names: Jin
+  orcid: "https://orcid.org/0009-0000-0539-5951"
+- family-names: Lalor
+  given-names: Edmund C.
+  orcid: "https://orcid.org/0000-0002-2498-6631"
+contact:
+- family-names: Bialas
+  given-names: Ole
+  orcid: "https://orcid.org/0000-0003-4472-7626"
+doi: 10.5281/zenodo.8321912
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Bialas
+    given-names: Ole
+    orcid: "https://orcid.org/0000-0003-4472-7626"
+  - family-names: Dou
+    given-names: Jin
+    orcid: "https://orcid.org/0009-0000-0539-5951"
+  - family-names: Lalor
+    given-names: Edmund C.
+    orcid: "https://orcid.org/0000-0002-2498-6631"
+  date-published: 2023-09-12
+  doi: 10.21105/joss.05657
+  issn: 2475-9066
+  issue: 89
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5657
+  title: "mTRFpy: A Python package for temporal response function
+    analysis"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05657"
+  volume: 8
+title: "mTRFpy: A Python package for temporal response function
+  analysis"
