### 2023-09-13 d900235
more informative error messages
